Adapt naming convention for collection.convert null safety test
Test classes not ending in "Test" are not executed in sbt. IntelliJ
runs them.

After this patch: 803 tests executed both in sbt and junit.
